feat: 스터디 관리하기 뷰 홀더 구현

+++ android/app/src/main/java/com/created/team201/presentation/studyManagement/model/StudyRoundDetailUiModel.kt
@@ -5,7 +5,7 @@ import com.created.team201.presentation.home.model.TodoUiModel
 
 data class StudyRoundDetailUiModel(
     val id: Long,
-    val masterId: Int,
+    val masterId: Long,
     val role: Role,
     val necessaryTodo: TodoUiModel,
     val optionalTodos: List<TodoUiModel>,
